Fix lamp's hitbox

+++ common/fr/mcnanotech/kevin_68/nanotechmod/main/client/renderer/RenderReinforcedFishingHook.java
@@ -1,7 +1,7 @@
 /**
  * This work is made available under the terms of the Creative Commons Attribution License:
  * http://creativecommons.org/licenses/by-nc-sa/4.0/deed.en
- * 
+ *
  * Cette œuvre est mise à disposition selon les termes de la Licence Creative Commons Attribution:
  * http://creativecommons.org/licenses/by-nc-sa/4.0/deed.fr
  */
@@ -37,10 +37,10 @@ public class RenderReinforcedFishingHook extends Render
         Tessellator tessellator = Tessellator.instance;
         byte b0 = 1;
         byte b1 = 2;
-        float f2 = (float)(b0 * 8 + 0) / 128.0F;
-        float f3 = (float)(b0 * 8 + 8) / 128.0F;
-        float f4 = (float)(b1 * 8 + 0) / 128.0F;
-        float f5 = (float)(b1 * 8 + 8) / 128.0F;
+        float f2 = (b0 * 8 + 0) / 128.0F;
+        float f3 = (b0 * 8 + 8) / 128.0F;
+        float f4 = (b1 * 8 + 0) / 128.0F;
+        float f5 = (b1 * 8 + 8) / 128.0F;
         float f6 = 1.0F;
         float f7 = 0.5F;
         float f8 = 0.5F;
@@ -48,10 +48,10 @@ public class RenderReinforcedFishingHook extends Render
         GL11.glRotatef(-this.renderManager.playerViewX, 1.0F, 0.0F, 0.0F);
         tessellator.startDrawingQuads();
         tessellator.setNormal(0.0F, 1.0F, 0.0F);
-        tessellator.addVertexWithUV((double)(0.0F - f7), (double)(0.0F - f8), 0.0D, (double)f2, (double)f5);
-        tessellator.addVertexWithUV((double)(f6 - f7), (double)(0.0F - f8), 0.0D, (double)f3, (double)f5);
-        tessellator.addVertexWithUV((double)(f6 - f7), (double)(1.0F - f8), 0.0D, (double)f3, (double)f4);
-        tessellator.addVertexWithUV((double)(0.0F - f7), (double)(1.0F - f8), 0.0D, (double)f2, (double)f4);
+        tessellator.addVertexWithUV(0.0F - f7, 0.0F - f8, 0.0D, f2, f5);
+        tessellator.addVertexWithUV(f6 - f7, 0.0F - f8, 0.0D, f3, f5);
+        tessellator.addVertexWithUV(f6 - f7, 1.0F - f8, 0.0D, f3, f4);
+        tessellator.addVertexWithUV(0.0F - f7, 1.0F - f8, 0.0D, f2, f4);
         tessellator.draw();
         GL11.glDisable(GL12.GL_RESCALE_NORMAL);
         GL11.glPopMatrix();
@@ -65,27 +65,27 @@ public class RenderReinforcedFishingHook extends Render
             vec3.rotateAroundY(-(entityFishHook.angler.prevRotationYaw + (entityFishHook.angler.rotationYaw - entityFishHook.angler.prevRotationYaw) * par9) * (float)Math.PI / 180.0F);
             vec3.rotateAroundY(f10 * 0.5F);
             vec3.rotateAroundX(-f10 * 0.7F);
-            double d3 = entityFishHook.angler.prevPosX + (entityFishHook.angler.posX - entityFishHook.angler.prevPosX) * (double)par9 + vec3.xCoord;
-            double d4 = entityFishHook.angler.prevPosY + (entityFishHook.angler.posY - entityFishHook.angler.prevPosY) * (double)par9 + vec3.yCoord;
-            double d5 = entityFishHook.angler.prevPosZ + (entityFishHook.angler.posZ - entityFishHook.angler.prevPosZ) * (double)par9 + vec3.zCoord;
+            double d3 = entityFishHook.angler.prevPosX + (entityFishHook.angler.posX - entityFishHook.angler.prevPosX) * par9 + vec3.xCoord;
+            double d4 = entityFishHook.angler.prevPosY + (entityFishHook.angler.posY - entityFishHook.angler.prevPosY) * par9 + vec3.yCoord;
+            double d5 = entityFishHook.angler.prevPosZ + (entityFishHook.angler.posZ - entityFishHook.angler.prevPosZ) * par9 + vec3.zCoord;
             double d6 = entityFishHook.angler == Minecraft.getMinecraft().thePlayer ? 0.0D : (double)entityFishHook.angler.getEyeHeight();
 
             if(this.renderManager.options.thirdPersonView > 0 || entityFishHook.angler != Minecraft.getMinecraft().thePlayer)
             {
                 float f11 = (entityFishHook.angler.prevRenderYawOffset + (entityFishHook.angler.renderYawOffset - entityFishHook.angler.prevRenderYawOffset) * par9) * (float)Math.PI / 180.0F;
-                double d7 = (double)MathHelper.sin(f11);
-                double d8 = (double)MathHelper.cos(f11);
-                d3 = entityFishHook.angler.prevPosX + (entityFishHook.angler.posX - entityFishHook.angler.prevPosX) * (double)par9 - d8 * 0.35D - d7 * 0.85D;
-                d4 = entityFishHook.angler.prevPosY + d6 + (entityFishHook.angler.posY - entityFishHook.angler.prevPosY) * (double)par9 - 0.45D;
-                d5 = entityFishHook.angler.prevPosZ + (entityFishHook.angler.posZ - entityFishHook.angler.prevPosZ) * (double)par9 - d7 * 0.35D + d8 * 0.85D;
+                double d7 = MathHelper.sin(f11);
+                double d8 = MathHelper.cos(f11);
+                d3 = entityFishHook.angler.prevPosX + (entityFishHook.angler.posX - entityFishHook.angler.prevPosX) * par9 - d8 * 0.35D - d7 * 0.85D;
+                d4 = entityFishHook.angler.prevPosY + d6 + (entityFishHook.angler.posY - entityFishHook.angler.prevPosY) * par9 - 0.45D;
+                d5 = entityFishHook.angler.prevPosZ + (entityFishHook.angler.posZ - entityFishHook.angler.prevPosZ) * par9 - d7 * 0.35D + d8 * 0.85D;
             }
 
-            double d9 = entityFishHook.prevPosX + (entityFishHook.posX - entityFishHook.prevPosX) * (double)par9;
-            double d10 = entityFishHook.prevPosY + (entityFishHook.posY - entityFishHook.prevPosY) * (double)par9 + 0.25D;
-            double d11 = entityFishHook.prevPosZ + (entityFishHook.posZ - entityFishHook.prevPosZ) * (double)par9;
-            double d12 = (double)((float)(d3 - d9));
-            double d13 = (double)((float)(d4 - d10));
-            double d14 = (double)((float)(d5 - d11));
+            double d9 = entityFishHook.prevPosX + (entityFishHook.posX - entityFishHook.prevPosX) * par9;
+            double d10 = entityFishHook.prevPosY + (entityFishHook.posY - entityFishHook.prevPosY) * par9 + 0.25D;
+            double d11 = entityFishHook.prevPosZ + (entityFishHook.posZ - entityFishHook.prevPosZ) * par9;
+            double d12 = ((float)(d3 - d9));
+            double d13 = ((float)(d4 - d10));
+            double d14 = ((float)(d5 - d11));
             GL11.glDisable(GL11.GL_TEXTURE_2D);
             GL11.glDisable(GL11.GL_LIGHTING);
             tessellator.startDrawing(3);
@@ -95,7 +95,7 @@ public class RenderReinforcedFishingHook extends Render
             for(int i = 0; i <= b2; ++i)
             {
                 float f12 = (float)i / (float)b2;
-                tessellator.addVertex(par2 + d12 * (double)f12, par4 + d13 * (double)(f12 * f12 + f12) * 0.5D + 0.25D, par6 + d14 * (double)f12);
+                tessellator.addVertex(par2 + d12 * f12, par4 + d13 * (f12 * f12 + f12) * 0.5D + 0.25D, par6 + d14 * f12);
             }
 
             tessellator.draw();
